samples/ml/ml_jobs/rl_finetune/requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
sglang>=0.2.0
snowflake-snowpark-python>=1.10.0
snowflake-ml-python>=1.4.0
